Rename github-auth to GitHubAuth
