+++ src/components/App/App.jsx
@@ -6,7 +6,7 @@ const Filter = require('../../models/Filter')
 const Filters = require('../../models/Filters')
 const GitHub = require('../../models/github')
 const AppMenu = require('../../models/AppMenu')
-const GitHubAuth = require('../../models/github-auth')
+const GitHubAuth = require('../../models/GitHubAuth')
 const LastFilter = require('../../models/last-filter')
 
 const TaskList = require('../TaskList')
