+++ src/models/GitHubAuth.js
